apps/api/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
pydantic==2.10.6
python-multipart==0.0.20
sqlalchemy==2.0.38
psycopg[binary]==3.2.5
pgvector==0.3.6
python-dotenv==1.0.1
httpx==0.28.1
google-genai==1.62.0